# BLD/PERF: bump cython to 0.29.19 for searchsorted