--- a/ci/deps/azure-36-minimum_versions.yaml
+++ b/ci/deps/azure-36-minimum_versions.yaml
@@ -5,7 +5,7 @@ dependencies:
   - python=3.6.1
 
   # tools
-  - cython=0.29.16
+  - cython=0.29.19
   - pytest=5.0.1
   - pytest-xdist>=1.21
   - hypothesis>=3.58.0