--- a/ci/deps/travis-36-slow.yaml
+++ b/ci/deps/travis-36-slow.yaml
@@ -6,7 +6,7 @@ dependencies:
   - python=3.6.*
 
   # tools
-  - cython>=0.29.16
+  - cython>=0.29.19
   - pytest>=5.0.1
   - pytest-xdist>=1.21
   - hypothesis>=3.58.0