--- a/ci/deps/travis-38.yaml
+++ b/ci/deps/travis-38.yaml
@@ -6,7 +6,7 @@ dependencies:
   - python=3.8.*
 
   # tools
-  - cython>=0.29.16
+  - cython>=0.29.19
   - pytest>=5.0.1
   - pytest-xdist>=1.21
   - hypothesis>=3.58.0