--- a/environment.yml
+++ b/environment.yml
@@ -12,7 +12,7 @@ dependencies:
   - asv
 
   # building
-  - cython>=0.29.16
+  - cython>=0.29.19
 
   # code checks
   - black=19.10b0