--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -5,7 +5,7 @@ numpy>=1.15
 python-dateutil>=2.7.3
 pytz
 asv
-cython>=0.29.16
+cython>=0.29.19
 black==19.10b0
 cpplint
 flake8<3.8.0