@jbrockmendel 2020-05-31
@scoder any idea how to get a more helpful traceback?

```
multiprocessing.pool.RemoteTraceback: 
"""
Traceback (most recent call last):
  File "/home/vsts/miniconda3/envs/pandas-dev/lib/python3.7/multiprocessing/pool.py", line 121, in worker
    result = (True, func(*args, **kwds))
  File "/home/vsts/miniconda3/envs/pandas-dev/lib/python3.7/multiprocessing/pool.py", line 44, in mapstar
    return list(map(*args))
  File "/home/vsts/miniconda3/envs/pandas-dev/lib/python3.7/site-packages/Cython/Build/Dependencies.py", line 1249, in cythonize_one_helper
    return cythonize_one(*m)
  File "/home/vsts/miniconda3/envs/pandas-dev/lib/python3.7/site-packages/Cython/Build/Dependencies.py", line 1225, in cythonize_one
    raise CompileError(None, pyx_file)
Cython.Compiler.Errors.CompileError: pandas/_libs/tslibs/conversion.pyx
"""

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "setup.py", line 792, in <module>
    setup_package()
  File "setup.py", line 762, in setup_package
    ext_modules=maybe_cythonize(extensions, compiler_directives=directives),
  File "setup.py", line 538, in maybe_cythonize
    return cythonize(extensions, *args, **kwargs)
  File "/home/vsts/miniconda3/envs/pandas-dev/lib/python3.7/site-packages/Cython/Build/Dependencies.py", line 1093, in cythonize
    result.get(99999)  # seconds
  File "/home/vsts/miniconda3/envs/pandas-dev/lib/python3.7/multiprocessing/pool.py", line 657, in get
    raise self._value
Cython.Compiler.Errors.CompileError: pandas/_libs/tslibs/conversion.pyx
```