@jbrockmendel 2020-06-01
No progress debugging this, we can wait for 3.0 anyway.  closing