@scoder 2020-06-02
@jbrockmendel the main part of that stack trace seems missing. There should be more output somewhere else in the build log, probably hidden by the parallel build. A serial build would also bring the output in order again.